The new on of hangman
In this version, I wrote 7 functions to deal with the word from a file. Each line in this file should contain two elements that are separated by a tab. One is the word at first, the other is the description of the word. Description can be any contents. In my example (test.txt), it contains the translation of the word in Chinese. The functions are as follows. The "count_lines" counts the number of line (or words) in the file. The "rand_num" generates a random integer from 0 to a maximum. The "get_line" gets a line from a file by the line number. The "get_word" separates the word by a certain separator. The "lower_word" transfers all alphabetic characters in a word to lower case. The "inser_word" inserts a type of separator between every two characters in a word. The "mask_word" replaces all alphabetic characters in a word to another character.