Add some new function.
In this version, I add 5 new functions about processing guess by user. The "check_input" to checks a character is legal. The "get_input" processes a alphabetic character in lower case. The "check_guess" checks if a character has guessed. The "make_guess" processes the actions after guessing a character. The "check_continue" checks the if this game continues.